test: cover stretch grips and drag regressions
